split and move files into shard module
