+++ lib/collection/src/operations/payload_ops.rs
@@ -5,7 +5,7 @@ use serde;
 use serde::{Deserialize, Serialize};
 use std::collections::HashMap;
 
-use crate::shard::ShardId;
+use crate::ShardId;
 
 use super::{split_iter_by_shard, OperationToShard, SplitByShard};
