+++ lib/collection/src/shard/local_shard.rs
@@ -33,93 +33,10 @@ use crate::operations::CollectionUpdateOperations;
 use crate::optimizers_builder::build_optimizers;
 use crate::update_handler::{OperationData, Optimizer, UpdateHandler, UpdateSignal};
 use crate::wal::SerdeWal;
-use crate::CollectionId;
+use crate::{CollectionId, ShardId};
 use segment::payload_storage::schema_storage::SchemaStorage;
 use segment::segment_constructor::load_segment;
 use std::fs::{read_dir, remove_dir_all};
-
-pub type ShardId = u32;
-
-pub type PeerId = u32;
-
-/// Shard
-///
-/// A shard can either be local or remote
-///
-pub enum Shard {
-    Local(LocalShard),
-    Remote(RemoteShard)
-}
-
-impl Shard {
-    pub async fn before_drop(&mut self) {
-        match self {
-            Shard::Local(local_shard) => local_shard.before_drop().await,
-            Shard::Remote(_) => (),
-        }
-    }
-
-    pub async fn update(
-        &self,
-        operation: CollectionUpdateOperations,
-        wait: bool,
-    ) -> CollectionResult<UpdateResult> {
-        match self {
-            Shard::Local(local_shard) => local_shard.update(operation, wait).await,
-            Shard::Remote(_) => todo!(),
-        }
-    }
-
-    pub fn segments(&self) -> &RwLock<SegmentHolder> {
-        match self {
-            Shard::Local(local_shard) => local_shard.segments(),
-            Shard::Remote(_) => todo!(),
-        }
-    }
-
-    pub async fn scroll_by(
-        &self,
-        segment_searcher: &(dyn CollectionSearcher + Sync),
-        offset: Option<ExtendedPointId>,
-        limit: usize,
-        with_payload_interface: &WithPayloadInterface,
-        with_vector: bool,
-        filter: Option<&Filter>,
-    ) -> CollectionResult<Vec<Record>> {
-        match self {
-            Shard::Local(local_shard) => local_shard.scroll_by(segment_searcher, offset, limit, with_payload_interface, with_vector, filter).await,
-            Shard::Remote(_) => todo!(),
-        }
-    }
-
-    pub async fn update_optimizer_params(
-        &self,
-        optimizer_config_diff: OptimizersConfigDiff,
-    ) -> CollectionResult<()> {
-        match self {
-            Shard::Local(local_shard) => local_shard.update_optimizer_params(optimizer_config_diff).await,
-            Shard::Remote(_) => todo!(),
-        }
-    }
-
-    pub async fn info(&self) -> CollectionResult<CollectionInfo> {
-        match self {
-            Shard::Local(local_shard) => local_shard.info().await,
-            Shard::Remote(_) => todo!(),
-        }
-    }
-}
-
-/// RemoteShard
-///
-/// Remote Shard is a representation of a shard that is located on a remote peer.
-/// Currently a placeholder implementation for later work.
-#[allow(dead_code)]
-pub struct RemoteShard {
-    id: ShardId,
-    collection_id: CollectionId,
-    peer_id: PeerId,
-}
 
 /// LocalShard
 ///
